+++ packages/fhir/test/shared/faker/fhir/elements/immunizationPractitioner.ts
@@ -1,9 +1,0 @@
-import { createMockFactory } from '@minvws/mgo-utils/test/shared';
-import { type ImmunizationPractitioner } from 'fhir/r3';
-import { reference } from '../type';
-
-export const immunizationPractitioner = createMockFactory<ImmunizationPractitioner>(() => {
-    return {
-        actor: reference(),
-    };
-});