+++ packages/fhir/test/shared/faker/fhir/type/period.ts
@@ -1,8 +1,0 @@
-import { createMockFactory } from '@minvws/mgo-utils/test/shared';
-import { type Period } from 'fhir/r3';
-import { dateTime } from './dateTime';
-
-export const period = createMockFactory<Period>(() => ({
-    start: dateTime(),
-    end: dateTime(),
-}));