+++ packages/fhir/test/shared/faker/fhirR3/elements/conditionStage.ts
@@ -1,0 +1,8 @@
+import { createMockFactory, mockArray } from '@minvws/mgo-utils/test/shared';
+import { type ConditionStage } from 'fhir/r3';
+import { codeableConcept, coding, reference } from '../type/index.js';
+
+export const conditionStage = createMockFactory<ConditionStage>(() => ({
+    summary: codeableConcept({ coding: mockArray({ max: 1, factory: coding }) }),
+    assessment: mockArray({ max: 5, factory: reference }),
+}));