+++ packages/fhir/test/shared/faker/fhirR3/elements/consentData.ts
@@ -1,0 +1,10 @@
+import { createMockFactory } from '@minvws/mgo-utils/test/shared';
+import { type ConsentData } from 'fhir/r3';
+import { code, reference } from '../type/index.js';
+
+export const consentData = createMockFactory<ConsentData>(() => {
+    return {
+        reference: reference(),
+        meaning: code(['instance', 'related', 'dependents', 'authoredby']),
+    };
+});