+++ packages/fhir/test/shared/faker/fhirR3/elements/medicationIngredient.ts
@@ -1,0 +1,13 @@
+import { faker } from '@faker-js/faker';
+import { createMockFactory } from '@minvws/mgo-utils/test/shared';
+import { type MedicationIngredient } from 'fhir/r3';
+import { codeableConcept, ratio, reference } from '../type/index.js';
+
+export const medicationIngredient = createMockFactory<MedicationIngredient>(() => {
+    return {
+        amount: ratio(),
+        isActive: faker.datatype.boolean(),
+        itemCodeableConcept: codeableConcept(),
+        itemReference: reference(),
+    };
+});