+++ packages/fhir/test/shared/faker/fhirR3/type/period.ts
@@ -1,0 +1,8 @@
+import { createMockFactory } from '@minvws/mgo-utils/test/shared';
+import { type Period } from 'fhir/r3';
+import { dateTime } from './dateTime.js';
+
+export const period = createMockFactory<Period>(() => ({
+    start: dateTime(),
+    end: dateTime(),
+}));