+++ packages/fhir/test/shared/faker/fhirR3/type/timing.ts
@@ -1,0 +1,14 @@
+import { createMockFactory, mockArray } from '@minvws/mgo-utils/test/shared';
+import { type Timing } from 'fhir/r3';
+import { codeableConcept } from './codeableConcept.js';
+import { dateTime } from './dateTime.js';
+import { timingRepeat } from './timingRepeat.js';
+
+export const timing = createMockFactory<Timing>(() => ({
+    code: codeableConcept(),
+    event: mockArray({
+        factory: dateTime,
+        max: 5,
+    }),
+    repeat: timingRepeat(),
+}));