+++ packages/hcim-ui/test/faker/index.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import { fhirFaker } from '@minvws/mgo-fhir/test/shared';
+import { fhirR3Faker } from '@minvws/mgo-fhir/test/shared';
 import { mgoFaker } from '@minvws/mgo-hcim-parse/test/shared';
 import { nullish } from '@minvws/mgo-utils/test/shared';
 import { uiFaker } from '../shared';
@@ -16,7 +16,7 @@ const custom: {
 type CustomizedFaker = typeof faker & {
     ui: typeof uiFaker;
     mgo: typeof mgoFaker;
-    fhir: typeof fhirFaker;
+    fhir: typeof fhirR3Faker;
     custom: typeof custom;
 };
 
@@ -24,7 +24,7 @@ const customizedFaker = faker as CustomizedFaker;
 
 customizedFaker.ui = uiFaker;
 customizedFaker.mgo = mgoFaker;
-customizedFaker.fhir = fhirFaker;
+customizedFaker.fhir = fhirR3Faker;
 customizedFaker.custom = custom;
 
 export { customizedFaker as faker };