+++ packages/hcim/test/faker/index.ts
@@ -1,12 +1,12 @@
 import { faker } from '@faker-js/faker';
-import { fhirFaker } from '@minvws/mgo-fhir/test/shared';
+import { fhirR3Faker } from '@minvws/mgo-fhir/test/shared';
 
 type CustomizedFaker = typeof faker & {
-    fhir: typeof fhirFaker;
+    fhir: typeof fhirR3Faker;
 };
 
 const customizedFaker = faker as CustomizedFaker;
 
-customizedFaker.fhir = fhirFaker;
+customizedFaker.fhir = fhirR3Faker;
 
 export { customizedFaker as faker };