+++ packages/utils/CHANGELOG.md
@@ -1,3 +1,35 @@
+## 0.1.1-10 (2025-07-24)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-9 (2025-07-24)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-8 (2025-07-24)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-7 (2025-07-23)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-6 (2025-07-23)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-5 (2025-07-23)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-4 (2025-07-23)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
+## 0.1.1-3 (2025-07-23)
+
+This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.
+
 ## 0.1.1-2 (2025-07-10)
 
 This was a version bump only for @minvws/mgo-utils to align it with other projects, there were no code changes.